Fixed Fifo issue
